refactor: world data
This commit introduces a new interface WorldData, and the old WorldData class is renamed to AllayWorldData and is moved to server module. The new world data interface won't contain any implementation-specific fields & methods.

We also removed most of the unused fields in old world data class, as we see that pmmp's world can still be loaded by vanilla correctly whose level.dat file only contains a small set of properties.

+++ api/src/main/java/org/allaymc/api/eventbus/event/world/TimeChangeEvent.java
@@ -11,16 +11,16 @@ import org.allaymc.api.world.WorldData;
  */
 @Getter
 public class TimeChangeEvent extends WorldEvent implements CancellableEvent {
-    protected long oldTime;
-    protected long newTime;
+    protected int oldTime;
+    protected int newTime;
 
-    public TimeChangeEvent(World world, long oldTime, long newTime) {
+    public TimeChangeEvent(World world, int oldTime, int newTime) {
         super(world);
         this.oldTime = oldTime;
         setNewTime(newTime);
     }
 
-    public void setNewTime(long time) {
+    public void setNewTime(int time) {
         Preconditions.checkArgument(time >= WorldData.TIME_DAY && time <= WorldData.TIME_FULL, "Time must be between " + WorldData.TIME_DAY + " and " + WorldData.TIME_FULL);
         this.newTime = time;
     }
